# openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: Zed Auth — Management API
  version: 0.1.0
  description: |
    The Management REST API for the Zed Auth service.

    **Every capability in the management console is available here.** The
    console is a client of this API and has no privileged path around it
    (`docs/PLAN/02` FR-14). If the console can do something this API cannot, that is
    a defect.

    ## Current status

    Phase 0. The service runs, and the endpoints below are the ones that exist:
    operational probes only. The identity and authorization endpoints arrive in
    Phase 1 and are deliberately absent from this document until they are real.

    ## Versioning

    Management endpoints are prefixed `/v1`. Breaking changes go to `/v2` with a
    published deprecation window for `/v1` (`docs/PLAN/05` Part B). The operational
    probes are unversioned: they describe the process, not the API.

    ## Authentication

    Management endpoints require an OAuth 2.0 Bearer access token, including
    the Client Credentials grant for service accounts. Authorization is
    enforced server-side on every request without exception — the console
    hiding a control is a user-experience decision, never a security one
    (`docs/PLAN/08`).
  contact:
    name: Responsible disclosure
    email: security@zedth.my.id
  license:
    # 3.0 has no `identifier` field, so the SPDX form used at 3.1 is not
    # available here. The name alone is unambiguous enough for its purpose: a
...
servers:
...
paths:
...
  /oauth/userinfo:
    get:
      tags: [OAuth]
      operationId: userinfo
      summary: UserInfo endpoint
      description: |
        Returns the claims the presented access token's scopes authorise, and
        nothing beyond them.

        **Send the token in the `Authorization` header.** RFC 6750 also defines
        an `access_token` query parameter and calls it NOT RECOMMENDED; this
        service does not accept one. A query parameter is written to the access
        log of every proxy in the path, to the browser's history, and to the
        `Referer` of whatever the page loads next.

        **Scopes decide the claims.** `openid` yields `sub`; `profile` yields
        `name`, `preferred_username` and `updated_at`; `email` yields `email`.
        A claim outside the granted scopes is never returned, and a claim whose
        underlying value is not set is omitted rather than returned empty.

        `email_verified` is **not** returned. Nothing in this service verifies
        an address yet, and asserting `false` would say we checked and it is
        not verified — which we did not.

        **`sub` is the user's identifier and is stable.** It is not sequential,
        not guessable, and never the email address. Store it as the user's
        permanent key; an address may change.

        **The token stops working the moment the session behind it ends.** An
        access token is not looked up, but the session it was issued from is —
        so a logout, a revocation, or a deactivated account invalidates it
        immediately rather than at its expiry.

        **Every unusable token gets the same answer**: `401` with
        `WWW-Authenticate: Bearer error="invalid_token"`. Expired, forged,
        wrong audience, and revoked are not distinguished, because a caller
        that could tell them apart could ask this endpoint whether a user has
        logged out.

        A token that is otherwise valid but lacks the `openid` scope gets `403`
        with `error="insufficient_scope"`, which is a different thing and is
        worth telling you: your token is fine, and it does not cover this.
      security:
        - oauth2: [openid]
      responses:
        "200":
          description: The claims the granted scopes authorise.
          headers:
            Cache-Control:
              description: Always `no-store`. The response is personal data addressed to one caller.
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
        "401":
          description: |
            The access token is missing, expired, forged, for another
            audience, or its session has ended. All of these are the same
            answer, deliberately.
          headers:
            WWW-Authenticate:
              description: |
                `Bearer realm="<issuer>"`, with
                `error="invalid_token"` when a credential was presented.
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: The token is valid and does not carry the `openid` scope.
          headers:
            WWW-Authenticate:
              description: '`Bearer error="insufficient_scope", scope="openid"`.'
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      description: |
        OAuth 2.0 access tokens. Authorization Code with PKCE for interactive
        clients including the console itself; Client Credentials for service
        accounts (`docs/PLAN/05` Part A).
      flows:
        authorizationCode:
          authorizationUrl: https://auth.zedth.my.id/oauth/authorize
          tokenUrl: https://auth.zedth.my.id/oauth/token
          refreshUrl: https://auth.zedth.my.id/oauth/token
          scopes:
            openid: OpenID Connect authentication.
            profile: Basic profile claims.
            email: Email address claim.
            offline_access: Issue a refresh token.
        clientCredentials:
          tokenUrl: https://auth.zedth.my.id/oauth/token
          scopes: {}
...
  schemas:
...
    ErrorCode:
      type: string
      description: |
        A stable, machine-readable code. Clients branch on this; `message` is
        for humans and may be reworded at any time without a version bump.

        Deliberately coarse. A code per failure mode leaks internal structure
        and becomes an enumeration oracle — `USER_NOT_FOUND` versus
        `PERMISSION_DENIED` tells an attacker which addresses are registered
        (`docs/SECURITY/02` §12).
      enum:
        - VALIDATION_ERROR
        - UNAUTHENTICATED
        - PERMISSION_DENIED
        - NOT_FOUND
        - CONFLICT
        - RATE_LIMITED
        - INTERNAL
    ErrorDetail:
      type: object
      description: |
        One specific problem within a failed request. Present for
        `VALIDATION_ERROR`, where naming the bad field is helpful and discloses
        nothing the caller did not send. Generally absent otherwise.
      required: [field, issue]
      additionalProperties: false
      properties:
        field:
          type: string
          description: JSON path to the offending field, relative to the request body.
          example: email
        issue:
          type: string
          description: What is wrong with it, in plain language.
          example: email format is not valid
...
    UserInfo:
      type: object
      description: |
        The claims the presented access token's scopes authorise.

        Only `sub` is guaranteed. Every other property appears when its scope
        was granted AND the underlying value is set — an absent claim means the
        value is not known or not authorised, and the two are deliberately not
        distinguished.
      required: [sub]
      properties:
        sub:
          type: string
          format: uuid
          description: |
            The user's stable identifier. Not sequential, not guessable, and
            never the email address — store this as the user's permanent key.
        name:
          type: string
          description: Display name. Requires the `profile` scope.
        preferred_username:
          type: string
          description: |
            Requires the `profile` scope. A user may change it, so it is not
            an identifier — use `sub`.
        updated_at:
          type: integer
          format: int64
          description: |
            Seconds since the epoch, when the user record last changed
            (OIDC Core 5.1). Requires the `profile` scope.
        email:
          type: string
          format: email
          description: |
            Requires the `email` scope. `email_verified` is deliberately not
            returned: nothing in this service verifies an address yet, and
            asserting `false` would claim a check that never happened.
...
    Error:
      type: object
      description: |
        The error envelope for every non-2xx response, without exception
        (`docs/PLAN/05` Part B § Standard Error Format). One shape means a client
        writes one error path rather than one per endpoint.
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              $ref: "#/components/schemas/ErrorCode"
            message:
              type: string
              description: |
                Human-readable, safe to surface in a UI, and never containing a
                token, a password, a raw `/v1/authz/check` resource attribute,
                a stack trace, or a SQL fragment (`docs/PLAN/13`, `CLAUDE.md`).
            details:
              type: array
              description: Field-level problems. Omitted when there are none.
              items:
                $ref: "#/components/schemas/ErrorDetail"
...
security:
  - oauth2: []